Make WorkDerivativeLoader provide mapping/hash-like interface, via proxy.
`NewspaperWorks::Data::WorkDerivativeLoader` is no longer an Enumerable
of mere destination name (e.g. extension) keys, but a mapping of such
names to path values.  Paths are loaded just-in-time, upon access of any
hash-like method or upon use of the `.path(name)`, `.with_io(name)`, or
`.data(name)` methods.  Callers are adjusted to this interface change.