Extracting convenience method
As part of my read through of this gem, I'm looking at noop refactors
that can help expose the expected public interface of the class
interactions.

This refactor follows on the principle of the Law of Demeter; namely try
to avoid method chains.  Method chains aren't intrinsicly bad, but
there's also a command/query separation to consider.

This change helps those that interact with
`NewspaperWorks::Data::WorkDerivatives` not need to know about the
implementation details of that class.  Instead exposing convenience
methods for that class.